[BUG_FIXED] Fix an infinite loop due to smart indent.
git-svn-id: svn://svn.tuxfamily.org/svnroot/notepadplus/repository/trunk@1297 f5eea248-9336-0410-98b8-ebc06183d4e3

+++ PowerEditor/src/Notepad_plus.cpp
@@ -2302,7 +2302,7 @@ int Notepad_plus::findMachedBracePos(size_t startPos, size_t endPos, char target
 
 	if (startPos > endPos) // backward
 	{
-		for (size_t i = startPos; i >= endPos; --i)
+		for (int i = startPos; i >= int(endPos); --i)
 		{
 			char aChar = (char)_pEditView->execute(SCI_GETCHARAT, i);
 			if (aChar == targetSymbol)
